Some semi-completed work from last year.
The last time I was working on this, I had done some hacking then
never got around to finishing it up.  This is that work in it's sort
of completed state.  This adds the rough concept of an Index and Task
object.

Still needs to be fleshed out some more:
* Tasks need to have a better concept of being loaded vs. being created
* Index should be able to perform lookup based on Subjects