Move the `yabt` script to scripts/ directory
Since `yabt` isn't really a binary, I'm moving it to the scripts/
directory to better reflect what it really is.  Reading through
the distutils documentation, it appears that this is the standard
for Python programs.

This commit also removes the hack to the sys.path that I had added
in in my earlier testing.